// src/type/type.ts
export interface ITodo {
    id: number
    title: string
    completed: boolean
}

export interface IAddress {
    street: string
    suite: string
    city: string
    zipcode: string
}
export interface ICompany {
    name: string
    catchPhrase: string
    bs: string
}

export interface IUser {
    id: number
    name: string
    username: string
    email: string
    address: IAddress
    phone: string
    website: string
    company: ICompany
}